refactor: `OAUTH_SCOPE_REGEX` を oauth-external に移動
