+++ workspaces/server/src/constants/oauth-external.ts
@@ -1,3 +1,6 @@
+export const OAUTH_SCOPE_REGEX =
+	/^[\x21|\x23-\x5B|\x5D-\x7E]+(?:\x20+[\x21|\x23-\x5B|\x5D-\x7E]+)*$/u;
+
 export const BAD_REQUEST_RESPONSE = [
 	"Bad Request",
 	400,
