+++ workspaces/server/src/routes/oauth/callback.ts
@@ -1,6 +1,6 @@
 import { vValidator } from "@hono/valibot-validator";
 import * as v from "valibot";
-import { OAUTH_SCOPE_REGEX } from "../../constants/oauth";
+import { OAUTH_SCOPE_REGEX } from "../../constants/oauth-external";
 import { SCOPE_IDS } from "../../constants/scope";
 import { factory } from "../../factory";
 import { cookieAuthMiddleware } from "../../middleware/auth";
